Explicit cast.

+++ src/storage/list.c
@@ -155,7 +155,7 @@ storageLstInsert(StorageList *const this, const unsigned int idx, const StorageI
             }
 
             case storageInfoLevelType:
-                listInfo.type.type = info->type;
+                listInfo.type.type = (uint8_t)info->type;
 
             default:
                 break;
